requirements.txt
# Core deep-learning stack
torch>=2.1
torchvision>=0.16

# Numerics + plotting
numpy>=1.24
matplotlib>=3.7
pillow>=10.0

# Notebook execution (only needed if you run the .ipynb via nbconvert)
jupyter>=1.0
ipykernel>=6.25
nbconvert>=7.7

# Dashboard / report generation
openpyxl>=3.1
python-docx>=1.1

# Interactive demo (only needed for `python app.py`)
gradio>=4.0

# Utilities
tqdm>=4.65
